Rename graphql package to match directory
